Split NSX-V related structs to separate file
